Use correct drawable when dragging newly created folder
When just creating a new folder, if user starts dragging it,
it uses the icon of the first item in the folder, instead of the
folder icon. When app is restarted it works as expected. The reason
it happened is that `drawable` is set to folder's drawable only when
initially loading from database.