Fix widgets getting stuck or disappearing
A Widget gridItem is created with widgetId -1.
When a Widget was binded for the first time, a
widgetId is generated by appWidgetHost.
However, this widgetId was not saved at DB.

During an onResume(), refreshLaunchers() would be
called, and then call fetchGridItems().
This would reset the widgetId to -1 and disconnect
it from the widgetId generated by appWidgetHost,
causing weird behaviors.

Another problem is that a new widgetId was being
generated everytime we replaced the Widget at a
restart process.

Changed so bindWidget() is only called when
binding the Widget for the first time, and then
we save the widgetId at the DB.

Later on, when recreating the screen, we retrieve
the widgetId from DB and call placeAppWidget() to
replace the Widget were it was.

Signed-off-by: Rafael Rondão <rafael.rondao@gmail.com>